string and logic controls

+++ src/main/kotlin/uniproc/internal/operators/math/MathPlusOperator.kt
@@ -56,7 +56,7 @@ class MathPlusOperator: Operator("+") {
             var output = ""
             // now try and combine them, prefering the more precise type
             if (leftSide.type==MathTypes.FLOAT.string || rightSide.type==MathTypes.FLOAT.string) {
-                output = (leftSide.value.toFloat() + rightSide.value.toFloat()).toString()
+                output = (leftSide.value.toDouble() + rightSide.value.toDouble()).toString()
             } else {
                 output = (leftSide.value.toInt() + rightSide.value.toInt()).toString()
             }
